requirements.txt
Flask==1.1.1
pylint==2.4.2
flask-sqlalchemy==2.4.1
pyjwt==1.7.1
pyyaml==5.1.2
elasticsearch==6.4.0
requests==2.22.0
python-dateutil
gunicorn==20.0.4
mysqlclient==1.4.2.post1
flask-cors==3.0.8
marshmallow==3.5.0